Refactor test so it checks status-code.
Previously, the test would use the presence or absence of
a request token to determine whether the response was synchronous or
asynchronous.  This was wrong and doesn't work for dCache, as it
always returns a token.

The test was also refactored to flow better, cover error cases better,
and to clean up only when the mkdir was successful.