Rewrite RmdirBeingPutInto usecase test
The existing code was buggy as it used the presence of the request token
to indicate whether a request was being processed asynchronously.  The
test had a number of other deficiencies, including poor cleanup and lack
of detail.

This patch rewrites the test in a cleaner fashion, robust cleanup and
clear information.