Fix PutOverwriteTransfParallel to not run using separate threads
The test checks how an SRM endpoint behaves when it receives two
srmPrepareToPut on the same SURL.

The test runs the two srmPrepareToPut on different threads, using S2's
support for parallel execution.

This is problematic because it introduces a race-condition in which
order the srmPrepareToPut requests are processed.  The exact order
depends on the scheduling of the client machine in addition to how the
server processes concurrent requests.