Fix tests to be robust against PrepareToPut being queued.
When a PrepareToPut request is queued, no TURL is returned.  This
caused the srmPrepareToPut operation to fail, as the reg-ex pattern
did not match the supplied fileStatuses.

This patch changes the prepareToPut operation and any subsequent
StatusOfPutRequest operation to capture the entire fileStatuses
and only extract the TURL once the prepareToPut / statusOfPutRequest
shows a SRM_SUCCESS outcome.