Test for prepareToPut timing out
Motivation:

The SRM system will timeout a request some time after the desired
lifetime has expired.  This is currently not tested.

Modification:

Add a test that checks that the PUT request remains valid for the
desired lifetime, at some later time, it is expired.

Result:

Expiry of PUT requests is now tested.