Add modify User-Agent and add ClientInfo HTTP headers
Motivation:

When diagnosing a problem discovered by S2, it is
useful to distinguish S2's SRM traffic for this test
from the traffic from other tests and other clients.

Modification:

gsoap provides no convenient way to inject custom
headers into the request.  In the end, the only
half-way reasonable approach would be to wrap the
existing function for handling headers with custom
code that reacts on a specific header.  In this
case, User-Agent.  Given the server always sends a
User-Agent header, the wrapper function is always
called.

This allows the code to modify the User-Agent field
(to indicate that the client is actually S2) along
with injecting the ClientInfo custom header.

The ClientInfo custom header value shows the test
name (without any .s2) along with the pid.

Additional information may be added by setting the
CLIENT_INFO environment variable.  The format should
be <key>=<value>;<key>=<value>;... (e.g.,
"Job=s2/254").  The <value> must be placed in
double-quotes if the value contains a space or
semicolon.

Result:

Target machine is able to distinguish between
different S2 tests, with the possibility to inject
additional information.