Fix race condition when testing space expirary
Motivation:

When a space expires dCache initially responds with the error indicating
that the space has expired.  Later on, dCache responds with
SRM_INVALID_REQUEST as the space token is no longer known to dCache.

This change in behaviour provides a race-condition: if the
CheckGarbageSpaceCollector test is unlucky then dCache garbage-collects
the space and returns SRM_INVALID_REQUEST, which (incorrectly) fails the
test.

Modification:

Add support for SRM_FAILURE + SRM_INVALID_REQUEST as a valid failure for
the srmPrepareToPut request.

Result:

The race condition no longer results in a test failing.