update function declarations to be sequential
Motivation:

There exists various include files that provide common functions.  These
currently define all such functions without any ';; ' prefix, which
means all but the first declaration will happen in parallel, on separate
threads.

There is some unidentified race-condition with the code having the
following stack-trace:

    #0  0x00000030c440b019 in pthread_cond_wait@@GLIBC_2.3.2 () from /lib64/libpthread.so.0
    #1  0x0000000000417467 in Process::eval ()
    #2  0x0000000000420bcc in Process::eval_subtree ()
    #3  0x0000000000416b7f in Process::eval_with_timeout ()
    #4  0x0000000000421047 in Process::eval_sequential_repeats ()
    #5  0x0000000000421485 in Process::eval_repeats ()
    #6  0x0000000000417321 in Process::eval ()
    #7  0x0000000000405846 in s2_run ()
    #8  0x0000000000407818 in main ()

This problem was observed in Jenkins with the usecase/MvIntoDir.s2 SRM
v2.2 test.  The s2 binary was found in this dead-lock state having sent
no SRM requests.  This strongly suggests that the problem lies within
the S2 code prior to the upload; i.e.,

    TEST $MATCH{(?P<srmTest>.*).s2$ ${0}}
       #include ../include/PutCycle.s2
       #include ../include/GetCycle.s2

Excluding each function's declaration body for brevity, these '#include'
statements expand into:

    TEST $MATCH{(?P<srmTest>.*).s2$ ${0}}
       DEFUN RemoveTrailing s : r
       DEFUN RemoveLeading s : r
       DEFUN car list : car cdr
       DEFUN FileSize f : r
       DEFUN DiskUsage f : r
       DEFUN StorageTypes f : r
       DEFUN Lifetimes f : r
       DEFUN FileTransPut fileToPut transferURL
       DEFUN DefTransferURLs f : t
       DEFUN PutCycle filesToPut sURLs : putRequestToken
       DEFUN FileTransGet fileToGet transferURL
       DEFUN DefFileStatuses f : t
       DEFUN GetCycle filesToGet sURLs : getRequestToken

Speculating, but it seems that having such a number of parallel tasks in
the same branch increases the risk of the above deadlock.

Modification:

As the SRM tests do not use parallel execution of nodes, this patch
introduces a work-around: make the declarations sequential.

Unfortunately, this is slightly tricky as the first node in the S2 tree
or any sub-tree is parsed differently: it must not have any ';;', '&&'
or '||' prefix.  Therefore, this patch also introduces a no-op (NOP)
operation as the first item of any '#include' section.

With this patch, the above expansion becomes:

    TEST $MATCH{(?P<srmTest>.*).s2$ ${0}}
       NOP
       ;; DEFUN RemoveTrailing s : r
       ;; DEFUN RemoveLeading s : r
       ;; DEFUN car list : car cdr
       ;; DEFUN FileSize f : r
       ;; DEFUN DiskUsage f : r
       ;; DEFUN StorageTypes f : r
       ;; DEFUN Lifetimes f : r
       ;; DEFUN FileTransPut fileToPut transferURL
       ;; DEFUN DefTransferURLs f : t
       ;; DEFUN PutCycle filesToPut sURLs : putRequestToken
       ;; DEFUN FileTransGet fileToGet transferURL
       ;; DEFUN DefFileStatuses f : t
       ;; DEFUN GetCycle filesToGet sURLs : getRequestToken

Note that the same expansion will also work if the '#include' statements
are the first items in the file, and not within a "TEST ..." operation,
as above.

Result:

Hopefully remove the trigger for a dead-lock bug in s2.