Have '#include' statements first.
Motivation:

Although not required, it is common that '#include' statements are the
first item in a file.  Currently the tests have '#include' statements
underneath a statement that defines the 'srmTest' variable.

Since these '#include' statement only defines functions, this convention
is unnecessary, causes some difficulty in understanding exactly how
these values are expanded, and looks like a work-around for some bug.

Modification:

Move the '#include' statements as the first non-comment, non-empty lines
in the file.

All SRM 2.2 files have been updated to follow this style.  The SRM v2.1
have not been modified as they are only of historical interest, if that.

Result:

Hopefully easier to read s2 tests.